erts: Break out hash routines to a separate module
These are a mess of `#define` and `#undef` and are very difficult
to read as part of utils.c. Break them out into a separate file to
ease refactoring.
