+++ erts/emulator/beam/erl_utils.h
@@ -24,6 +24,7 @@
 #include "sys.h"
 #include "atom.h"
 #include "erl_printf.h"
+#include "erl_term_hashing.h"
 
 struct process;
 
@@ -63,17 +64,11 @@ erts_current_interval_acqb(erts_interval_t *icp)
  */
 void erts_silence_warn_unused_result(long unused);
 
-
 int erts_fit_in_bits_int64(Sint64);
 int erts_fit_in_bits_int32(Sint32);
 int erts_fit_in_bits_uint(Uint);
 Sint erts_list_length(Eterm);
 int erts_is_builtin(Eterm, Eterm, int);
-Uint32 make_hash2(Eterm);
-Uint32 trapping_make_hash2(Eterm, Eterm*, struct process*);
-Uint32 make_hash(Eterm);
-Uint32 make_internal_hash(Eterm, Uint32 salt);
-Uint32 make_map_hash(Eterm key, int level);
 
 void erts_save_emu_args(int argc, char **argv);
 Eterm erts_get_emu_args(struct process *c_p);
